// src/models/shared/productType.ts
export enum ProductType {
  LAPTOP = "Laptop",
  SMARTPHONE = "Smartphone",
  TELEVISION = "Television",
  FURNITURE = "Furniture",
  CAMERA = "Camera",
  CLOTHING = "Clothing",
  BOOK = "Book",
  SPORTS_EQUIPMENT = "Sports Equipment",
  TOY = "Toy",
  JEWELRY = "Jewelry",
  ART_SUPPLIES = "Art Supplies",
  AUTOMOTIVE_PART = "Automotive Part",
  MUSICAL_INSTRUMENT = "Musical Instrument",
  OFFICE_EQUIPMENT = "Office Equipment"
  }